remove chunk_copy_test.go
